+++ zebra-rpc/src/methods/trees.rs
@@ -8,13 +8,13 @@ use zebra_chain::{
     subtree::{NoteCommitmentSubtreeData, NoteCommitmentSubtreeIndex},
 };
 
-/// A subtree data type that can hold Sapling or Orchard subtree roots.
+/// A subtree data type that can hold Sapling, Orchard, or Ironwood subtree roots.
 pub type SubtreeRpcData = NoteCommitmentSubtreeData<String>;
 
 /// Response to a `z_getsubtreesbyindex` RPC request.
 ///
-/// Contains the Sapling or Orchard pool label, the index of the first subtree in the list,
-/// and a list of subtree roots and end heights.
+/// Contains the shielded pool label, the index of the first subtree in the
+/// list, and a list of subtree roots and end heights.
 #[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize, Getters, new)]
 pub struct GetSubtreesByIndexResponse {
     /// The shielded pool to which the subtrees belong.
@@ -28,7 +28,8 @@ pub struct GetSubtreesByIndexResponse {
 
     /// A sequential list of complete subtrees, in `index` order.
     ///
-    /// The generic subtree root type is a hex-encoded Sapling or Orchard subtree root string.
+    /// The generic subtree root type is a hex-encoded shielded subtree root
+    /// string.
     //
     // TODO: is this needed?
     //#[serde(skip_serializing_if = "Vec::is_empty")]
@@ -38,7 +39,7 @@ pub struct GetSubtreesByIndexResponse {
 impl Default for GetSubtreesByIndexResponse {
     fn default() -> Self {
         Self {
-            pool: "sapling | orchard".to_string(),
+            pool: "sapling | orchard | ironwood".to_string(),
             start_index: NoteCommitmentSubtreeIndex(u16::default()),
             subtrees: vec![],
         }
@@ -47,7 +48,7 @@ impl Default for GetSubtreesByIndexResponse {
 
 /// Response to a `z_gettreestate` RPC request.
 ///
-/// Contains hex-encoded Sapling & Orchard note commitment trees and their corresponding
+/// Contains hex-encoded shielded note commitment trees and their corresponding
 /// [`struct@Hash`], [`Height`], and block time.
 ///
 /// The format of the serialized trees represents `CommitmentTree`s from the crate
@@ -60,10 +61,12 @@ impl Default for GetSubtreesByIndexResponse {
 /// whereas in `CommitmentTree`, the vector of ommers is sparse with [`None`] values in the gaps.
 ///
 /// The dense format might be used in future RPCs.
-#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize, Getters, new)]
+///
+/// The serialized response omits `ironwood` unless Ironwood tree state is
+/// available for the requested block.
+#[derive(Clone, Debug, Eq, PartialEq, Getters)]
 pub struct GetTreestateResponse {
     /// The block hash corresponding to the treestate, hex-encoded.
-    #[serde(with = "hex")]
     #[getter(copy)]
     hash: Hash,
 
@@ -80,7 +83,6 @@ pub struct GetTreestateResponse {
     /// A treestate containing a Sprout note commitment tree, hex-encoded. Zebra
     /// does not support returning it; but the field is here to enable parsing
     /// responses from other implementations.
-    #[serde(skip_serializing_if = "Option::is_none")]
     sprout: Option<Treestate>,
 
     /// A treestate containing a Sapling note commitment tree, hex-encoded.
@@ -88,9 +90,79 @@ pub struct GetTreestateResponse {
 
     /// A treestate containing an Orchard note commitment tree, hex-encoded.
     orchard: Treestate,
+
+    /// A treestate containing an Ironwood note commitment tree, hex-encoded.
+    /// Serialized only when [`Self::has_ironwood`] returns true.
+    #[getter(skip)]
+    ironwood: Treestate,
+
+    /// Whether the serialized RPC response should include Ironwood tree state.
+    #[getter(skip)]
+    has_ironwood: bool,
 }
 
 impl GetTreestateResponse {
+    /// Constructs a new [`GetTreestateResponse`] with Ironwood tree state.
+    pub fn new(
+        hash: Hash,
+        height: Height,
+        time: u32,
+        sprout: Option<Treestate>,
+        sapling: Treestate,
+        orchard: Treestate,
+        ironwood: Treestate,
+    ) -> Self {
+        Self {
+            hash,
+            height,
+            time,
+            sprout,
+            sapling,
+            orchard,
+            ironwood,
+            has_ironwood: true,
+        }
+    }
+
+    /// Constructs a new [`GetTreestateResponse`] that may omit Ironwood tree state.
+    pub fn new_with_optional_ironwood(
+        hash: Hash,
+        height: Height,
+        time: u32,
+        sprout: Option<Treestate>,
+        sapling: Treestate,
+        orchard: Treestate,
+        ironwood: Option<Treestate>,
+    ) -> Self {
+        let has_ironwood = ironwood.is_some();
+
+        Self {
+            hash,
+            height,
+            time,
+            sprout,
+            sapling,
+            orchard,
+            ironwood: ironwood.unwrap_or_default(),
+            has_ironwood,
+        }
+    }
+
+    /// Returns the Ironwood treestate.
+    pub fn ironwood(&self) -> &Treestate {
+        &self.ironwood
+    }
+
+    /// Returns whether the Ironwood treestate was present in the RPC response.
+    pub fn has_ironwood(&self) -> bool {
+        self.has_ironwood
+    }
+
+    /// Returns the Ironwood treestate if it was present in the RPC response.
+    pub fn optional_ironwood(&self) -> Option<&Treestate> {
+        self.has_ironwood.then_some(&self.ironwood)
+    }
+
     /// Constructs [`Treestate`] from its constituent parts.
     #[deprecated(note = "Use `new` instead.")]
     pub fn from_parts(
@@ -112,7 +184,6 @@ impl GetTreestateResponse {
                 final_root: None,
             },
         };
-
         Self {
             hash,
             height,
@@ -120,6 +191,8 @@ impl GetTreestateResponse {
             sprout: None,
             sapling,
             orchard,
+            ironwood: Treestate::default(),
+            has_ironwood: false,
         }
     }
 
@@ -145,7 +218,78 @@ impl Default for GetTreestateResponse {
             sprout: Default::default(),
             sapling: Default::default(),
             orchard: Default::default(),
+            ironwood: Default::default(),
+            has_ironwood: false,
         }
+    }
+}
+
+impl serde::Serialize for GetTreestateResponse {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        #[derive(serde::Serialize)]
+        struct GetTreestateResponseRef<'a> {
+            #[serde(with = "hex")]
+            hash: Hash,
+            height: Height,
+            time: u32,
+            #[serde(skip_serializing_if = "Option::is_none")]
+            sprout: Option<&'a Treestate>,
+            sapling: &'a Treestate,
+            orchard: &'a Treestate,
+            #[serde(skip_serializing_if = "Option::is_none")]
+            ironwood: Option<&'a Treestate>,
+        }
+
+        let response = GetTreestateResponseRef {
+            hash: self.hash,
+            height: self.height,
+            time: self.time,
+            sprout: self.sprout.as_ref(),
+            sapling: &self.sapling,
+            orchard: &self.orchard,
+            ironwood: self.optional_ironwood(),
+        };
+
+        serde::Serialize::serialize(&response, serializer)
+    }
+}
+
+impl<'de> serde::Deserialize<'de> for GetTreestateResponse {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        #[derive(serde::Deserialize)]
+        struct GetTreestateResponseHelper {
+            #[serde(with = "hex")]
+            hash: Hash,
+            height: Height,
+            time: u32,
+            #[serde(default)]
+            sprout: Option<Treestate>,
+            sapling: Treestate,
+            orchard: Treestate,
+            #[serde(default)]
+            ironwood: Option<Treestate>,
+        }
+
+        let response =
+            <GetTreestateResponseHelper as serde::Deserialize>::deserialize(deserializer)?;
+        let has_ironwood = response.ironwood.is_some();
+
+        Ok(Self {
+            hash: response.hash,
+            height: response.height,
+            time: response.time,
+            sprout: response.sprout,
+            sapling: response.sapling,
+            orchard: response.orchard,
+            ironwood: response.ironwood.unwrap_or_default(),
+            has_ironwood,
+        })
     }
 }
 
@@ -154,7 +298,7 @@ impl Default for GetTreestateResponse {
 /// [1]: https://zcash.github.io/rpc/z_gettreestate.html
 #[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize, Getters, new)]
 pub struct Treestate {
-    /// Contains an Orchard or Sapling serialized note commitment tree,
+    /// Contains a shielded serialized note commitment tree,
     /// hex-encoded.
     commitments: Commitments,
 }
@@ -178,21 +322,20 @@ impl Default for Treestate {
     }
 }
 
-/// A wrapper that contains either an Orchard or Sapling note commitment tree.
+/// A wrapper that contains a shielded note commitment tree.
 ///
-/// Note that in the original [`z_gettreestate`][1] RPC, [`Commitments`] also
-/// contains the field `finalRoot`. Zebra does *not* use this field.
+/// `finalRoot` and `finalState` are omitted when a specific tree state is not
+/// available.
 ///
-/// [1]: https://zcash.github.io/rpc/z_gettreestate.html
 #[serde_with::serde_as]
 #[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize, Getters, new)]
 pub struct Commitments {
-    /// Orchard or Sapling serialized note commitment tree root, hex-encoded.
+    /// Shielded serialized note commitment tree root, hex-encoded.
     #[serde_as(as = "Option<serde_with::hex::Hex>")]
     #[serde(skip_serializing_if = "Option::is_none")]
     #[serde(rename = "finalRoot")]
     final_root: Option<Vec<u8>>,
-    /// Orchard or Sapling serialized note commitment tree, hex-encoded.
+    /// Shielded serialized note commitment tree, hex-encoded.
     #[serde_as(as = "Option<serde_with::hex::Hex>")]
     #[serde(skip_serializing_if = "Option::is_none")]
     #[serde(rename = "finalState")]